Rename scheduleMedicationReminders.js to scheduleReminderNotifications.js
